Assignments/Assignment4/requirements.txt
facenet_pytorch==2.6.0
matplotlib==3.8.4
opencv_python==4.9.0.80
pandas==2.2.2
Pillow>=10.2.0,<10.3.0
torch>=2.2.0,<2.3.0
tqdm==4.66.4
